backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# PDF & Documents
pypdf==5.1.0
Pillow>=10.0.0
reportlab==4.2.5
python-pptx==0.6.23
hwp-hwpx-parser>=1.0.0
python-hwpx>=2.8.0  # HWP 출력 시 필요

# AI
google-generativeai==0.8.3

# Utils
python-multipart==0.0.17
pydantic==2.10.2
pydantic-settings==2.6.1

# Async & Jobs
httpx==0.28.1
beautifulsoup4>=4.12.0
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
